+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 1.1 (unreleased)
 
+* Raise error when settings file exists, but is not readable
+
 * Added `settings_environment` option. It defaults to `Rails.env`.
 
   Example: